Some unittests for algo.py and util.py (#76)
* Added unittest for utils.py and algo.py

+++ python-sdk/nuscenes/eval/detection/utils.py
@@ -214,6 +214,6 @@ def cummean(x: np.array) -> np.array:
         return np.ones(len(x))  # If all errors are NaN set to error to 1 for all operating points.
     else:
         # Accumulate in a nan-aware manner.
-        sum_vals = np.nancumsum(x)  # Cumulative sum ignoring nans.
+        sum_vals = np.nancumsum(x.astype(float))  # Cumulative sum ignoring nans.
         count_vals = np.cumsum(~np.isnan(x))  # Number of non-nans up to each position.
         return np.divide(sum_vals, count_vals, out=np.zeros_like(sum_vals), where=count_vals != 0)
